feat: introduce context middleware
