+++ app/V5/Modules/Dashboard/Controllers/DashboardV5Controller.php
@@ -835,7 +835,7 @@ class DashboardV5Controller extends AppBaseController
 
     private function getCurrentSchoolId(Request $request): ?int
     {
-        // School ID is set by SchoolContextMiddleware in the request (same as seasons)
+        // School ID is set by ContextMiddleware in the request
         $schoolId = $request->get('context_school_id');
         if ($schoolId) {
             return (int) $schoolId;
@@ -861,18 +861,23 @@ class DashboardV5Controller extends AppBaseController
 
     private function getCurrentSeasonId(Request $request): ?int
     {
-        // Get season ID from header (sent by frontend interceptor)
+        // Context middleware sets the season id
+        $seasonId = $request->get('context_season_id');
+        if ($seasonId) {
+            return (int) $seasonId;
+        }
+
+        // Fallbacks for legacy headers/params
         $seasonId = $request->header('X-Season-ID');
         if ($seasonId) {
             return (int) $seasonId;
         }
-        
-        // Fallback to parameter
+
         $seasonId = $request->get('season_id');
         if ($seasonId) {
             return (int) $seasonId;
         }
-        
+
         return null; // Don't provide a default, let the method handle the null
     }
 }
